# apps/api/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "snapdone Go API. 인증 credential은 JWT가 아닌 opaque 서버 세션 토큰.",
        "title": "snapdone API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/processing-jobs": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "사진 한 장(multipart 필드 image)으로 처리 작업을 만들고 바로 돌아온다. 결과는 작업 조회로 받는다.\nJPEG · PNG · GIF · WebP만 받고 7,500,000 byte까지다. 사진은 저장하지 않는다.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processing"
                ],
                "summary": "사진 처리 시작",
                "parameters": [
                    {
                        "type": "file",
                        "description": "사진 한 장",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ProcessingJobResponse"
                        }
                    },
                    "400": {
                        "description": "multipart가 아니거나 image 필드 없음 (invalid_image)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "credential 없음 · 만료 · 취소 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "상한 초과 (image_too_large)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "415": {
                        "description": "지원하지 않는 형식 (unsupported_image)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "처리 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/processing-jobs/{jobId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "내 처리 작업의 상태와, 완료됐다면 결과. 다른 사용자의 작업은 없는 작업과 같다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "processing"
                ],
                "summary": "사진 처리 작업",
                "parameters": [
                    {
                        "type": "string",
                        "description": "처리 시작이 돌려준 jobId",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ProcessingJobResponse"
                        }
                    },
                    "401": {
                        "description": "credential 없음 · 만료 · 취소 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "없는 작업 (job_not_found)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "처리 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "httpserver.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "enum": [
                        "session_expired",
                        "provider_unavailable",
                        "invalid_callback",
                        "invalid_image",
                        "image_too_large",
                        "unsupported_image",
                        "job_not_found",
                        "invalid_onboarding",
                        "onboarding_complete",
                        "onboarding_out_of_order"
                    ],
                    "example": "session_expired"
                }
            }
        },
...
        "httpserver.ProcessingFactResponse": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string",
                    "example": "날짜"
                },
                "value": {
                    "type": "string",
                    "example": "8월 20일 19시"
                }
            }
        },
        "httpserver.ProcessingJobResponse": {
            "type": "object",
            "properties": {
                "jobId": {
                    "type": "string",
                    "example": "4f1c2a9e-0000-4000-8000-000000000000"
                },
                "result": {
                    "$ref": "#/definitions/httpserver.ProcessingResultResponse"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "running",
                        "completed",
                        "failed"
                    ],
                    "example": "running"
                }
            }
        },
        "httpserver.ProcessingResultResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "enum": [
                        "place",
                        "event",
                        "receipt",
                        "foreign_text",
                        "shopping",
                        "work",
                        "other"
                    ],
                    "example": "event"
                },
                "confidence": {
                    "type": "string",
                    "enum": [
                        "high",
                        "medium",
                        "low"
                    ],
                    "example": "high"
                },
                "facts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpserver.ProcessingFactResponse"
                    }
                },
                "suggestedAction": {
                    "type": "string",
                    "enum": [
                        "save_place",
                        "add_to_calendar",
                        "record_expense",
                        "translate",
                        "none"
                    ],
                    "example": "add_to_calendar"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003copaque session credential\u003e\" 형식. 서버만 해석할 수 있는 세션 토큰.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}